Remove test-setup script from main test script (#2245)
The `test-setup` script negatively affects testing because it always
tries installing the most recent types. This makes `npm test` fail for
older versions, which can be especially annoying while trying to `git
bisect`. The behaviour of testing the current compiler with the newest
types is retained in the github CI workflow, which calls the individual
test steps separately, and continues to execute `test-setup`.