Rework expressionChain.ts utils (#2242)
Improves the emit of
https://roblox-ts.com/playground/#code/AYEg3gjAvgBOBMtwGYlgCxWAbgLACgg
which is currently quite ugly due to unnecessary wrapping
```lua
-- Compiled with roblox-ts v2.0.4
local _ = tostring(1) .. (" " .. (tostring(2) .. (" " .. (tostring(3) .. (" " .. (tostring(4) .. (" " .. tostring(5))))))))
```

All operators currently passed to `binaryExpressionChain` are
commutative (*1), meaning the change does not affect runtime outcome.

For future code, should this function handle non-commutative operators,
and if so, how? I'd say that there's no single logical result of the
function for non-commutative operators, so it should just not support
them as parameter.

(*1) assuming they are not made impure using metamethods. If they are,
the behaviour after this change probably makes more sense.

---------

Co-authored-by: Osyris <osyrisrblx@gmail.com>