Pin exact versions of @roblox-ts dependencies (#2724)
Releasing luau-ast@1.1.0 broke roblox-ts for some users because it's
installed as `^1.0.1` in the compiler.

This release probably should have been 2.0.0 because we removed some
functions. I've deprecated 1.1.0 and published 2.0.0 instead.

There's little upside to using `^` because it's very unlikely users will
have multiple dependencies which use any of the @roblox-ts packages.

Instead, we will opt to use the `=` prefix to prevent this issue from
happening in the future.