Fix transformPaths transformer (#2800)
`transformPaths` broke because `resolver.moduleExportsSomeValue()` no
longer exists in TypeScript (and is not trivial to copy from an old
version).

Originally, I tried to just `npm install` the latest version, but this
had a few issues:
- TypeScript wasn't able to find types because of how the package.json
is setup, I was able to hack around this with a `// @ts-ignore`
- Newest version seems to not respect just `baseUrl`, it requires you
have a `paths` in your tsconfig.json (and many roblox-ts projects don't)

I also can't just copy over the latest file from the repo because it has
a bunch of imports.

Instead, I found a slightly newer version which doesn't use
`moduleExportsSomeValue` and is still a single file. This required a few
small fixes i.e. removing non-ts.factory APIs.

---

The first commit in this PR is copying the new version of the
transformer

The second (+ subsequent) commit is the required fixes.