Add back `yargs.fail()` (#2812)
This was removed in
https://github.com/roblox-ts/roblox-ts/commit/553259333b450a284a5f6bce7349ed36bc3e353b
because we thought it was effectively the same as the default behavior.

One key difference we missed: custom .fail() uses `console.log()` which
automatically calls `.toString()` on the error. For custom error classes
like CLIError, this is where we do the formatting for the output.

I've added it back + added a clarifying comment.